cleaned back end code and improved file structure, with dedicated router and controller files
